+++ packages/@expo/cli/src/start/server/metro/MetroBundlerDevServer.ts
@@ -646,6 +646,7 @@ export class MetroBundlerDevServer extends BundlerDevServer {
       parsedOptions,
       {
         isExporting: !!options.isExporting,
+        enableAtlas: options.enableAtlas,
         exp,
       }
     );